feat: removed ollama refs in chat and settings, moved some composable files to the correct locations
