+++ src/composables/useAIAssistant.ts
